nse_backend/requirements.txt
Django>=5.0,<5.1
django-cors-headers>=4.4.0
djangorestframework>=3.15.0
mysqlclient>=2.2.4
pandas
numpy
yfinance
scikit-learn
scipy
python-telegram-bot>=21.0
requests
